firmware/gateway/requirements.txt
paho-mqtt==1.6.1
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.1
RPi.GPIO==0.7.1; platform_system == "Linux"
adafruit-circuitpython-dht==4.0.2; platform_system == "Linux"
adafruit-blinka==8.25.0; platform_system == "Linux"
fake-rpi==0.7.1; platform_system == "Windows"